pyproject.toml
[tool.poetry]
name = "mapigen"
version = "0.1.0"
description = "Metadata API Generator - fetch, normalize, and compress OpenAPI specs"
authors = ["Evan Huang <meoya.dev@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "mapigen", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
niquests = {extras = ["h3", "brotli", "zstd"], version = "^3.15.2"}
pydantic = "^2.11.7"
jsonschema = "^4.25.1"
lz4 = "^4.3"
openapi-spec-validator = "^0.7.2"
openapi-schema-validator = "^0.6.3"
structlog = "^25.4.0"
msgspec = "^0.19.0"
zstandard = "^0.24.0"
tqdm = "^4.67.1"
psutil = "^7.0.0"
referencing = "0.36.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-asyncio = "1.1.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
snakeviz = "^2.2.2"
pytest-profiling = "^1.8.1"
pytest-sugar = "^1.0.0"
jinja2 = "^3.1.4"
